i32

+++ client.js
@@ -25,7 +25,7 @@ const client = minecraft.createClient({
 for (let entry of fake.entries()) {
   client.registerChannel(entry[1], [
     "container",
-    [{ name: "id", type: "i64" }, { name: "data", type: "string" }]
+    [{ name: "id", type: "i32" }, { name: "data", type: "string" }]
   ]);
 }
